Move `UserCx` to co-trap
