+++ mizu/lib/ksc-core/src/lib.rs
@@ -1,0 +1,9 @@
+#![cfg_attr(not(test), no_std)]
+#![feature(macro_metavar_expr)]
+
+extern crate alloc;
+
+mod error;
+mod raw_reg;
+
+pub use self::{error::*, raw_reg::*};
