+++ mizu/lib/ksc/Cargo.toml
@@ -10,11 +10,10 @@ test = ["dep:smol"]
 [dependencies]
 # Local crates
 co-trap = {path = "../co-trap"}
+ksc-core = {path = "../ksc-core"}
 # External crates
 ahash = {version = "0", default-features = false}
 bevy_utils_proc_macros = "0"
-enum-primitive-derive = "0"
 futures-util = {version = "0", default-features = false, features = ["alloc"]}
 hashbrown = {version = "0", default-features = false, features = ["inline-more"]}
-num-traits = {version = "0", default-features = false}
 smol = {version = "1", optional = true}
