+++ mizu/lib/ksc/src/lib.rs
@@ -1,10 +1,9 @@
 #![cfg_attr(not(test), no_std)]
-#![feature(macro_metavar_expr)]
 
 extern crate alloc;
 
-mod error;
 mod handler;
-mod raw_reg;
 
-pub use self::{error::*, handler::*, raw_reg::*};
+pub use ksc_core::*;
+
+pub use self::handler::*;
